requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
nltk==3.7
oauthlib==3.2.1
passlib==1.7.4
psycopg2-binary==2.9.4
python-dotenv==0.21.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
textblob==0.17.1
tqdm==4.64.1
twython==3.9.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
